+++ fimber_io/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.4.2] - Fixed versioning
+
+- fixed versioning
+
+## [0.4.1] - Bug fixes
+
+- [\#75 issue](https://github.com/magillus/flutter-fimber/issues/75) - Fix for TimedRollingFileTree missing passed logLevels and logFormat
+- [\#73 issue](https://github.com/magillus/flutter-fimber/issues/73) - Fix for mute levels list not being distinct
+
 ## 0.4.0 Fimber dependency on dart::io separated
 
 - moved all File based and dart::io package dependent code from `fimber` to `fimber_io`